+++ src/components/booking/overview/PerformanceOverview.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import OverviewBox from '@/components/overview/OverviewBox.vue';
+import OverviewBox from './OverviewBox.vue';
 
 export default {
   name: 'performance-overview-box',
