+++ src/components/booking/overview/UserOverview.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import OverviewBox from '@/components/overview/OverviewBox.vue';
+import OverviewBox from './OverviewBox.vue';
 
 export default {
   name: 'user-overview',
